requirements.txt
googletrans==3.0.0
langchain==0.2.12
langchain_google_genai==1.0.8
langdetect==1.0.9
protobuf==5.27.3
python-dotenv==1.0.1
pyttsx3==2.90
SpeechRecognition==3.10.4
